cleanup and final fixes
